7.x.x Refactoring
