+++ api/src/main/java/org/antublue/test/engine/api/TestEngine.java
@@ -95,11 +95,6 @@ public @interface TestEngine {
     @Retention(RetentionPolicy.RUNTIME)
     @interface Disabled {}
 
-    /** Store annotation */
-    @Target({ElementType.ANNOTATION_TYPE, ElementType.FIELD})
-    @Retention(RetentionPolicy.RUNTIME)
-    @interface Context {}
-
     /** Tag annotation */
     @Target({ElementType.ANNOTATION_TYPE, ElementType.TYPE, ElementType.METHOD})
     @Retention(RetentionPolicy.RUNTIME)
