+++ plugin/src/main/java/org/antublue/test/engine/maven/plugin/TestEngineMavenPlugin.java
@@ -68,7 +68,6 @@ public class TestEngineMavenPlugin extends AbstractMojo {
     private static final String VERSION = Information.getInstance().getVersion();
 
     static {
-        // Load the Store singleton in the current classloader
         Context.getInstance();
     }
